+++ app/src/main/java/bertrand/myopengl/CubeGray.java
@@ -44,7 +44,7 @@ public final class CubeGray extends ObjectModel {
                        -1, -1, -1,    0.5f, 0.5f, 0.5f, 1,     0, -1, 0, // 23
                 };
 
-                final byte[] indices  = {
+                final int[] indices  = {
                         // Front
                         0, 1, 2,
                         2, 3, 0,
